Onboarding: Create Deployment screen and make separation relative to status
