+++ src/components/Onboarding/Deployment.js
@@ -1,0 +1,7 @@
+import React from 'react'
+
+function Deployment() {
+  return <div>Deploy</div>
+}
+
+export default Deployment
