+++ src/components/Onboarding/statuses.js
@@ -1,0 +1,2 @@
+export const STATUS_GARDEN_SETUP = Symbol('STATUS_GARDEN_SETUP')
+export const STATUS_GARDEN_DEPLOYMENT = Symbol('STATUS_GARDEN_DEPLOYMENT')
